+++ app/javascript/components/charts/cp-performance-all-sectors/index.js
@@ -1,0 +1,3 @@
+import CPPerformanceAllSectors from './CPPerformanceAllSectors';
+
+export default CPPerformanceAllSectors;
