+++ app/javascript/components/charts/cp-performance/CPPerformance.js
@@ -1,3 +1,5 @@
+/* eslint-disable react/no-this-in-sfc */
+
 import React, { useEffect, useMemo, useState } from 'react';
 import { renderToString } from 'react-dom/server';
 import PropTypes from 'prop-types';
@@ -5,81 +7,9 @@ import PropTypes from 'prop-types';
 import Highcharts from 'highcharts';
 import HighchartsReact from 'highcharts-react-official';
 
-function createSVGGroup(className) {
-  const group = createSVGElement('g');
-  group.setAttribute('class', className);
-  group.setAttribute('opacity', 0.3);
-  return group;
-}
-
-function createSVGElement(element) {
-  return document.createElementNS('http://www.w3.org/2000/svg', element);
-}
-
-function Line({ dashStyle, color, width }) {
-  const style = {
-    borderBottomStyle: dashStyle === 'dot' ? 'dotted' : 'solid',
-    borderBottomColor: color,
-    width
-  };
-
-  return (
-    <span className="line" style={style} />
-  );
-}
-
-function LegendItem({ name, onRemove }) {
-  return (
-    <div className="legend-item">
-      <span className="legend-item__color" />
-      {name}
-      <button className="legend-item__remove" type="button" onClick={onRemove}>
-        x
-      </button>
-    </div>
-  );
-}
-
-LegendItem.propTypes = {
-  name: PropTypes.string.isRequired,
-  onRemove: PropTypes.func.isRequired
-};
-
-function SharedTooltip({ xValue, yValues, unit }) {
-  const companyValues = yValues.filter(v => !v.isBenchmark);
-  const benchmarkValues = yValues.filter(v => v.isBenchmark);
-
-  return (
-    <div className="cp-tooltip">
-      <div className="cp-tooltip__row cp-tooltip__row--bold">
-        <span>{xValue}</span>
-        <span>{unit}</span>
-      </div>
-
-      {companyValues.map(y => (
-        <div className="cp-tooltip__row cp-tooltip__row--bold">
-          <span className="cp-tooltip__value-title">
-            <Line dashStyle={y.dashStyle} color={y.color} />
-            {y.title}
-            {y.isTargeted && (<small>(targeted)</small>)}
-          </span>
-          <span>{y.value}</span>
-        </div>
-      ))}
-
-      <div className="cp-tooltip__targets">
-        TARGETS
-      </div>
-
-      {benchmarkValues.map(y => (
-        <div className="cp-tooltip__row">
-          <span>{y.title}</span>
-          <span>{y.value}</span>
-        </div>
-      ))}
-    </div>
-  );
-}
+import { createSVGGroup } from './helpers';
+import LegendItem from './LegendItem';
+import Tooltip from './Tooltip';
 
 function CPPerformanceAllSectors({ dataUrl, unit }) {
   const [data, setData] = useState([]);
@@ -187,13 +117,13 @@ function CPPerformanceAllSectors({ dataUrl, unit }) {
         }));
 
         return renderToString(
-          <SharedTooltip xValue={xValue} yValues={yValues} unit={unit} />
+          <Tooltip xValue={xValue} yValues={yValues} unit={unit} />
         );
       }
     },
     series: chartData
   };
-  const chartStyle = {
+  const chartWrapperStyle = {
     width: '100%',
     height: '800px'
   };
@@ -208,7 +138,7 @@ function CPPerformanceAllSectors({ dataUrl, unit }) {
         <span className="line-legend-item line-legend-item--reported">Reported</span>
         <span className="line-legend-item line-legend-item--targeted">Targeted</span>
       </div>
-      <div style={chartStyle}>
+      <div style={chartWrapperStyle}>
         <HighchartsReact
           highcharts={Highcharts}
           options={options}
