+++ app/javascript/components/charts/cp-performance/index.js
@@ -1,0 +1,3 @@
+import CPPerformance from './CPPerformance';
+
+export default CPPerformance;
