+++ app/views/tpi/sectors/show.html.erb
@@ -79,7 +79,7 @@
   <section id="carbon-performance" class="container">
     <h4>Carbon Performance: <%= @sector.name %></h4>
 
-    <%= react_component('charts/CPPerformance', {
+    <%= react_component('charts/cp-performance', {
       dataUrl: emissions_chart_data_tpi_sector_path(@sector.id),
       unit: @sector.latest_cp_unit&.unit
     }) %>
